refactor(view-engine): extract analysis editor labels and choice atoms
